DSP Jit more minor fixes
git-svn-id: https://dolphin-emu.googlecode.com/svn/trunk@5405 8ced0084-cf51-0410-be5f-012b33b47a6e

+++ Source/Core/DSPCore/Src/DSPHWInterface.cpp
@@ -125,7 +125,7 @@ u16 gdsp_mbox_read_l(u8 mbx)
 }
 
 
-void gdsp_ifx_write(u16 addr, u16 val)
+void gdsp_ifx_write(u32 addr, u32 val)
 {
 	switch (addr & 0xff)
 	{
